src/utils/schemas.py
from pydantic import BaseModel, ConfigDict

'''
pydantic
'''
class UserCreate(BaseModel):
    name: str
    email: str
    password: str

class UserLogin(BaseModel):
    username: str
    password: str

class UserRead(BaseModel):
    id: int
    name: str
    email: str

    model_config = ConfigDict(from_attributes = True)